Final project version

+++ Internfinder-BE/GAO_BackEnd/src/main/java/gao/internfinder/backend/repository/AccountRepository.java
@@ -1,14 +1,15 @@
 package gao.internfinder.backend.repository;
 
 
-import gao.internfinder.backend.Entity.EntityTrang.Account;
+
+import gao.internfinder.backend.Entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, Integer> {
 	Optional<Account> findByUsername(String username);
 
 	Boolean existsByUsername(String username);
